# IEXSYS/src/main/java/org/exp/iexsys/dto/RegisterRequest.java
package org.exp.iexsys.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public class RegisterRequest {

    @NotBlank(message = "用户名不能为空")
    @Size(min = 2, max = 16, message = "用户名长度需在2-16之间")
    private String username;

    @NotBlank(message = "密码不能为空")
    @Size(min = 6, max = 24, message = "密码长度需在6-24之间")
    private String password;

    private String realName;

    @Email(message = "邮箱格式不正确")
    private String email;

    private String phone;

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getRealName() {
        return realName;
    }

    public void setRealName(String realName) {
        this.realName = realName;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPhone() {
        return phone;
    }

    public void setPhone(String phone) {
        this.phone = phone;
    }
}