# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Online Exam System API",
    "version": "1.0.0",
    "description": "基于 online_exam_system.sql 数据模型和现有 auth 实现的 REST API 草案。认证采用会话 Cookie（JSESSIONID），响应使用统一结构 ApiResponse<T>{code,message,data}。"
  },
  "servers": [
    { "url": "/api" }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "JSESSIONID" }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "message": { "type": "string" },
          "data": {}
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" },
          "realName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "用户注册（已实现）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
          }
        },
        "responses": { "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "用户名密码登录（已实现）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": { "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
...
    "/auth/bind-phone": {
      "post": {
        "tags": ["Auth"],
        "summary": "绑定手机号/第三方",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone", "code"],
                "properties": {
                  "phone": { "type": "string" },
                  "code": { "type": "string" },
                  "wxOpenId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "ok" } }
      }
    },
...
  }
}